Attempt to make the message building of Twitch and IRC a bit nicer
